@@ -3,21 +3,20 @@
 # found in the LICENSE file.
 
 name: process_runner
-version: 3.1.0
+version: 4.0.0-nullsafety
 description: A process invocation astraction for Dart that manages a multiprocess queue.
 homepage: https://github.com/google/process_runner
 
 dependencies:
-  file: ^5.0.0
-  meta: ^1.1.2
-  path: ^1.5.1
-  platform: ^2.2.0
-  process: ^3.0.13
-  async: ^2.4.2
+  file: ^6.0.0-nullsafety.1
+  meta: ^1.3.0-nullsafety.2
+  path: ^1.8.0-nullsafety
+  platform: ^3.0.0-nullsafety.1
+  process: ^4.0.0-nullsafety.1
+  async: ^2.5.0-nullsafety
 
 dev_dependencies:
-  test: ^1.0.0
-  mockito: ^4.1.1
+  test: ^1.16.0-nullsafety.1
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.10.0-4.0.dev <2.10.0'
