@@ -9,6 +9,8 @@
 // of single-threaded CPU-intensive commands by a multple of the number of
 // processor cores you have (modulo being disk/network bound, of course).
 
+// @dart = 2.10
+
 import 'dart:io';
 
 import 'package:process_runner/process_runner.dart';
@@ -89,7 +91,7 @@ main.dart [flags]
 ''';
 }
 
-String findOption(String option, List<String> args) {
+String? findOption(String option, List<String> args) {
   for (int i = 0; i < args.length - 1; ++i) {
     if (args[i] == option) {
       return args[i + 1];
@@ -117,7 +119,7 @@ Future<void> main(List<String> args) async {
   final bool printReport = args.contains('--report');
   // If the numWorkers is set to null, then the ProcessPool will automatically
   // select the number of processes based on how many CPU cores the machine has.
-  final int numWorkers = int.tryParse(findOption('workers', args) ?? '');
+  final int? numWorkers = int.tryParse(findOption('workers', args) ?? '');
   final Directory workingDirectory = Directory(findOption('workingDirectory', args) ?? '.');
   final List<String> cmds = findAllOptions('cmd', args).toList();
 
@@ -126,7 +128,7 @@ Future<void> main(List<String> args) async {
   List<String> fileCommands = <String>[];
   // Read from stdin if the --file option is set to '-'.
   if (commandFile == '-') {
-    String line = stdin.readLineSync();
+    String? line = stdin.readLineSync();
     while (line != null) {
       fileCommands.add(line);
       line = stdin.readLineSync();
@@ -165,6 +167,6 @@ Future<void> main(List<String> args) async {
     if (printReport) {
       print('\nFinished job ${done.name}');
     }
-    stdout.write(done.result.stdout);
+    stdout.write(done.result!.stdout);
   }
 }
